split resolve-search-permission
